server/requirements.txt
psycopg2-binary
flask
flask_cors
